requirements.txt
# Core LangGraph / LangChain
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0

# FastAPI monitoring dashboard
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# WebSocket connectivity (Binance / Alpaca)
websockets>=12.0

# Technical-indicator computation
numpy>=1.26.0
pandas>=2.2.0
pandas-ta>=0.3.14b

# HTTP client (Exa.ai sentiment, REST calls)
httpx>=0.27.0

# Async support utilities
anyio>=4.4.0

# Environment / config
python-dotenv>=1.0.0

# Serialisation + settings validation
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Alpaca paper / live order execution (optional)
alpaca-py>=0.30.0
